refactor(wallet): split actions into modules
