+++ src/stores/wallet.ts
@@ -12,6 +12,9 @@ import { usePRStore } from "./payment-request";
 import { useWorkersStore } from "./workers";
 import { useInvoicesWorkerStore } from "./invoicesWorker";
 import { DEFAULT_BUCKET_ID } from "./buckets";
+import { invoiceActions } from "./wallet/invoices";
+import { tokenActions } from "./wallet/tokens";
+import { p2pkActions } from "./wallet/p2pk";
 
 import * as _ from "underscore";
 import token from "src/js/token";
@@ -177,6 +180,9 @@ export const useWalletStore = defineStore("wallet", {
     },
   },
   actions: {
+    ...invoiceActions,
+    ...tokenActions,
+    ...p2pkActions,
     mintWallet(url: string, unit: string): CashuWallet {
       // short-lived wallet for mint operations
       // note: the unit of the wallet will be activeUnit by default,
@@ -294,72 +300,6 @@ export const useWalletStore = defineStore("wallet", {
       }
       return chunks;
     },
-    coinSelectSpendBase64: function (
-      proofs: WalletProof[],
-      amount: number,
-    ): WalletProof[] {
-      const base64Proofs = proofs.filter((p) => !p.id.startsWith("00"));
-      if (base64Proofs.length > 0) {
-        base64Proofs.sort((a, b) => b.amount - a.amount);
-        let sum = 0;
-        let selectedProofs: WalletProof[] = [];
-        for (let i = 0; i < base64Proofs.length; i++) {
-          const proof = base64Proofs[i];
-          sum += proof.amount;
-          selectedProofs.push(proof);
-          if (sum >= amount) {
-            return selectedProofs;
-          }
-        }
-        return [];
-      }
-      return [];
-    },
-    coinSelect: function (
-      proofs: WalletProof[],
-      wallet: CashuWallet,
-      amount: number,
-      includeFees: boolean = false,
-      preferredBucketId?: string,
-    ): WalletProof[] {
-      if (proofs.reduce((s, t) => (s += t.amount), 0) < amount) {
-        // there are not enough proofs to pay the amount
-        return [];
-      }
-
-      let orderedProofs = [...proofs];
-
-      if (preferredBucketId) {
-        orderedProofs.sort((a, b) => {
-          const aPref = a.bucketId === preferredBucketId ? 0 : 1;
-          const bPref = b.bucketId === preferredBucketId ? 0 : 1;
-          if (aPref !== bPref) {
-            return aPref - bPref;
-          }
-          return b.amount - a.amount;
-        });
-      }
-
-      let sum = 0;
-      let selectedProofs: WalletProof[] = [];
-      for (const proof of orderedProofs) {
-        selectedProofs.push(proof);
-        sum += proof.amount;
-        const fees = includeFees ? wallet.getFeesForProofs(selectedProofs) : 0;
-        if (sum >= amount + fees) {
-          break;
-        }
-      }
-
-      const finalFees = includeFees
-        ? wallet.getFeesForProofs(selectedProofs)
-        : 0;
-      if (sum < amount + finalFees) {
-        return [];
-      }
-
-      return selectedProofs.map((p) => ({ ...p, reserved: false } as WalletProof));
-    },
     spendableProofs: function (proofs: WalletProof[], amount: number) {
       const uIStore = useUiStore();
       const proofsStore = useProofsStore();
@@ -381,362 +321,6 @@ export const useWalletStore = defineStore("wallet", {
     },
     getFeesForProofs: function (proofs: Proof[]): number {
       return this.wallet.getFeesForProofs(proofs);
-    },
-    sendToLock: async function (
-      proofs: WalletProof[],
-      wallet: CashuWallet,
-      amount: number,
-      receiverPubkey: string,
-      bucketId: string = DEFAULT_BUCKET_ID,
-      locktime?: number,
-      refundPubkey?: string,
-    ) {
-      const mintStore = useMintsStore();
-      const nutSupport = mintStore.activeInfo?.nut_supports || [];
-      if (!(nutSupport.includes(10) && nutSupport.includes(11))) {
-        notifyError(this.t("wallet.notifications.lock_not_supported"));
-        throw new Error("Mint does not support timelocks or P2PK");
-      }
-      const spendableProofs = this.spendableProofs(proofs, amount);
-      const proofsToSend = this.coinSelect(
-        spendableProofs,
-        wallet,
-        amount,
-        true,
-        bucketId,
-      );
-      const keysetId = this.getKeyset(wallet.mint.mintUrl, wallet.unit);
-      const { keep: keepProofs, send: sendProofs } = await wallet.send(
-        amount,
-        proofsToSend,
-        { keysetId, pubkey: receiverPubkey, locktime, refund: refundPubkey },
-      );
-      const proofsStore = useProofsStore();
-      await proofsStore.removeProofs(proofsToSend);
-      // note: we do not store sendProofs in the proofs store but
-      // expect from the caller to store it in the history
-      await proofsStore.addProofs(keepProofs, undefined, bucketId, "");
-      return { keepProofs, sendProofs };
-    },
-    send: async function (
-      proofs: WalletProof[],
-      wallet: CashuWallet,
-      amount: number,
-      invalidate: boolean = false,
-      includeFees: boolean = false,
-      bucketId: string = DEFAULT_BUCKET_ID,
-    ): Promise<{ keepProofs: Proof[]; sendProofs: Proof[] }> {
-      /*
-      splits proofs so the user can keep firstProofs, send scndProofs.
-      then sets scndProofs as reserved.
-
-      if invalidate, scndProofs (the one to send) are invalidated
-      */
-      const mintStore = useMintsStore();
-      const proofsStore = useProofsStore();
-      const uIStore = useUiStore();
-      let proofsToSend: WalletProof[] = [];
-      const keysetId = this.getKeyset(wallet.mint.mintUrl, wallet.unit);
-      await uIStore.lockMutex();
-      try {
-        const spendableProofs = this.spendableProofs(proofs, amount);
-
-        proofsToSend = this.coinSelect(
-          spendableProofs,
-          wallet,
-          amount,
-          includeFees,
-          bucketId,
-        );
-        const totalAmount = proofsToSend.reduce((s, t) => (s += t.amount), 0);
-        const fees = includeFees ? wallet.getFeesForProofs(proofsToSend) : 0;
-        const targetAmount = amount + fees;
-
-        let keepProofs: Proof[] = [];
-        let sendProofs: Proof[] = [];
-
-        if (totalAmount != targetAmount) {
-          const counter = this.keysetCounter(keysetId);
-          proofsToSend = this.coinSelect(
-            spendableProofs,
-            wallet,
-            targetAmount,
-            true,
-            bucketId,
-          );
-          ({ keep: keepProofs, send: sendProofs } = await wallet.send(
-            targetAmount,
-            proofsToSend,
-            { counter, keysetId, proofsWeHave: spendableProofs },
-          ));
-          this.increaseKeysetCounter(
-            keysetId,
-            keepProofs.length + sendProofs.length,
-          );
-          await proofsStore.addProofs(keepProofs, undefined, bucketId, "");
-          await proofsStore.addProofs(sendProofs, undefined, bucketId, "");
-
-          // make sure we don't delete any proofs that were returned
-          const proofsToSendNotReturned = proofsToSend
-            .filter((p) => !sendProofs.find((s) => s.secret === p.secret))
-            .filter((p) => !keepProofs.find((k) => k.secret === p.secret));
-          await proofsStore.removeProofs(proofsToSendNotReturned);
-        } else if (totalAmount == targetAmount) {
-          keepProofs = [];
-          sendProofs = proofsToSend;
-        } else {
-          throw new Error("could not split proofs.");
-        }
-
-        await proofsStore.setReserved(sendProofs, true);
-        if (invalidate) {
-          await proofsStore.removeProofs(sendProofs);
-        }
-        return { keepProofs, sendProofs };
-      } catch (error: any) {
-        await proofsStore.setReserved(proofsToSend, false);
-        console.error(error);
-        notifyApiError(error);
-        this.handleOutputsHaveAlreadyBeenSignedError(keysetId, error);
-        throw error;
-      } finally {
-        uIStore.unlockMutex();
-      }
-    },
-    /**
-     *
-     *
-     * @param {array} proofs
-     */
-    redeem: async function (bucketId: string = DEFAULT_BUCKET_ID) {
-      /*
-      Receives a token that is prepared in the receiveToken – it is not yet in the history
-      */
-      const uIStore = useUiStore();
-      const mintStore = useMintsStore();
-      const p2pkStore = useP2PKStore();
-
-      receiveStore.showReceiveTokens = false;
-
-      if (receiveStore.receiveData.tokensBase64.length == 0) {
-        throw new Error("no tokens provided.");
-      }
-      const tokenJson = token.decode(receiveStore.receiveData.tokensBase64);
-      if (tokenJson == undefined) {
-        throw new Error("no tokens provided.");
-      }
-      let proofs = token.getProofs(tokenJson);
-      if (proofs.length == 0) {
-        throw new Error("no proofs found.");
-      }
-      const inputAmount = proofs.reduce((s, t) => (s += t.amount), 0);
-      let fee = 0;
-      let mintInToken = token.getMint(tokenJson);
-      let unitInToken = token.getUnit(tokenJson);
-
-      const historyToken = {
-        amount: inputAmount,
-        token: receiveStore.receiveData.tokensBase64,
-        unit: unitInToken,
-        mint: mintInToken,
-        label: receiveStore.receiveData.label ?? "",
-        fee: fee,
-        bucketId,
-      } as HistoryToken;
-      const mintWallet = this.mintWallet(historyToken.mint, historyToken.unit);
-      const mint = mintStore.mints.find((m) => m.url === historyToken.mint);
-      if (!mint) {
-        throw new Error("mint not found");
-      }
-      await uIStore.lockMutex();
-      try {
-        // redeem
-        const keysetId = this.getKeyset(historyToken.mint, historyToken.unit);
-        const counter = this.keysetCounter(keysetId);
-        const privkey = receiveStore.receiveData.p2pkPrivateKey;
-        let proofs: Proof[];
-        try {
-          proofs = await mintWallet.receive(
-            receiveStore.receiveData.tokensBase64,
-            {
-              counter,
-              privkey,
-              proofsWeHave: mintStore.mintUnitProofs(mint, historyToken.unit),
-            },
-          );
-          await proofsStore.addProofs(
-            proofs,
-            undefined,
-            bucketId,
-            receiveStore.receiveData.label ?? "",
-          );
-          this.increaseKeysetCounter(keysetId, proofs.length);
-        } catch (error: any) {
-          console.error(error);
-          this.handleOutputsHaveAlreadyBeenSignedError(keysetId, error);
-          throw new Error("Error receiving tokens: " + error);
-        }
-
-        p2pkStore.setPrivateKeyUsed(privkey);
-
-        const outputAmount = proofs.reduce((s, t) => (s += t.amount), 0);
-
-        // if token is already in history, set to paid, else add to history
-        if (
-          tokenStore.historyTokens.find(
-            (t) =>
-              t.token === receiveStore.receiveData.tokensBase64 && t.amount > 0,
-          )
-        ) {
-          tokenStore.setTokenPaid(receiveStore.receiveData.tokensBase64);
-        } else {
-          // if this is a self-sent token, we will find an outgoing token with the inverse amount
-          if (
-            tokenStore.historyTokens.find(
-              (t) =>
-                t.token === receiveStore.receiveData.tokensBase64 &&
-                t.amount < 0,
-            )
-          ) {
-            tokenStore.setTokenPaid(receiveStore.receiveData.tokensBase64);
-          }
-          fee = inputAmount - outputAmount;
-          historyToken.fee = fee;
-          historyToken.amount = outputAmount;
-          tokenStore.addPaidToken(historyToken);
-        }
-        useUiStore().vibrate();
-        let message = this.t("wallet.notifications.received", {
-          amount: uIStore.formatCurrency(outputAmount, historyToken.unit),
-        });
-        if (fee > 0) {
-          message += this.t("wallet.notifications.fee", {
-            fee: uIStore.formatCurrency(fee, historyToken.unit),
-          });
-        }
-        notifySuccess(message);
-      } catch (error: any) {
-        console.error(error);
-        notifyApiError(error);
-        throw error;
-      } finally {
-        uIStore.unlockMutex();
-      }
-      // }
-    },
-
-    // /mint
-    /**
-     * Ask the mint to generate an invoice for the given amount
-     * Upon paying the request, the mint will credit the wallet with
-     * cashu tokens
-     */
-    requestMint: async function (
-      amount: number,
-      mintWallet: CashuWallet,
-    ): Promise<MintQuoteResponse> {
-      try {
-        // create MintQuotePayload(this.invoiceData.amount) payload
-        const payload: MintQuotePayload = {
-          amount: amount,
-          unit: mintWallet.unit,
-        };
-        const data = await mintWallet.mint.createMintQuote(payload);
-        this.invoiceData.amount = amount;
-        this.invoiceData.bolt11 = data.request;
-        this.invoiceData.quote = data.quote;
-        this.invoiceData.date = currentDateStr();
-        this.invoiceData.status = "pending";
-        this.invoiceData.mint = mintWallet.mint.mintUrl;
-        this.invoiceData.unit = mintWallet.unit;
-        this.invoiceData.mintQuote = data;
-        this.invoiceHistory.push({
-          ...this.invoiceData,
-        });
-        return data;
-      } catch (error: any) {
-        console.error(error);
-        notifyApiError(
-          error,
-          this.t("wallet.notifications.could_not_request_mint"),
-        );
-        throw error;
-      } finally {
-      }
-    },
-    mint: async function (
-      invoice: InvoiceHistory,
-      verbose: boolean = true,
-      bucketId: string = DEFAULT_BUCKET_ID,
-    ) {
-      const proofsStore = useProofsStore();
-      const mintStore = useMintsStore();
-      const tokenStore = useTokensStore();
-      const uIStore = useUiStore();
-      const keysetId = this.getKeyset(invoice.mint, invoice.unit);
-      const mintWallet = this.mintWallet(invoice.mint, invoice.unit);
-      const mint = mintStore.mints.find((m) => m.url === invoice.mint);
-      if (!mint) {
-        throw new Error("mint not found");
-      }
-
-      await uIStore.lockMutex();
-      try {
-        // first we check if the mint quote is paid
-        const mintQuote = await mintWallet.checkMintQuote(invoice.quote);
-        invoice.mintQuote = mintQuote;
-        console.log("### mint(): mintQuote", mintQuote);
-        switch (mintQuote.state) {
-          case MintQuoteState.PAID:
-            break;
-          case MintQuoteState.UNPAID:
-            if (verbose) {
-              notify(this.t("wallet.notifications.invoice_still_pending"));
-            }
-            throw new Error("invoice pending.");
-          case MintQuoteState.ISSUED:
-            throw new Error("invoice already issued.");
-          default:
-            throw new Error("unknown state.");
-        }
-        // MintQuoteState must be PAID
-        const counter = this.keysetCounter(keysetId);
-        const proofs = await mintWallet.mintProofs(
-          invoice.amount,
-          invoice.quote,
-          {
-            keysetId,
-            counter,
-            proofsWeHave: mintStore.mintUnitProofs(mint, invoice.unit),
-          },
-        );
-        this.increaseKeysetCounter(keysetId, proofs.length);
-        await proofsStore.addProofs(proofs, undefined, bucketId, "");
-
-        // update UI
-        await this.setInvoicePaid(invoice.quote);
-        const serializedProofs = proofsStore.serializeProofs(proofs);
-        tokenStore.addPaidToken({
-          amount: invoice.amount,
-          token: serializedProofs,
-          unit: invoice.unit,
-          mint: invoice.mint,
-          label: "",
-          bucketId,
-        });
-        useInvoicesWorkerStore().removeInvoiceFromChecker(invoice.quote);
-
-        return proofs;
-      } catch (error: any) {
-        console.error(error);
-        if (verbose) {
-          notifyApiError(error);
-        }
-        this.handleOutputsHaveAlreadyBeenSignedError(keysetId, error);
-        throw error;
-      } finally {
-        uIStore.unlockMutex();
-      }
     },
     // get a melt quote for the current invoice data
     meltQuoteInvoiceData: async function () {
